chore: refactor validateRoleId middleware and deleteRole controller

+++ src/controllers/role/deleteRole.ts
@@ -4,6 +4,7 @@ import validateRoleId from "../../middleware/role/validateRoleId";
 import {
   sendResponse,
   sendServerError,
+  STATUS_FORBIDDEN,
   STATUS_OK,
 } from "../../utilities/response";
 
@@ -11,7 +12,15 @@ const deleteRole = async (req: Request, res: Response) => {
   const prisma = new PrismaClient();
 
   try {
-    const role = res.locals.role;
+    const { role } = res.locals;
+
+    if (role.People.length > 0) {
+      sendResponse(res, STATUS_FORBIDDEN, {
+        message: "Role is assigned to some people",
+      });
+
+      return;
+    }
 
     await prisma.role.delete({
       where: {
@@ -22,6 +31,8 @@ const deleteRole = async (req: Request, res: Response) => {
     sendResponse(res, STATUS_OK, {
       message: "Role deleted successfully",
     });
+
+    return;
   } catch (error) {
     sendServerError(res, error);
   } finally {
